feat: manage listings

+++ app/Http/Controllers/ListingController.php
@@ -46,7 +46,7 @@ class ListingController extends Controller
     }
 
     $formFields['user_id'] = auth()->id();
-    
+
     Listing::create($formFields);
     return redirect("/")->with('message', 'Listing successfully created');
   }
@@ -57,6 +57,10 @@ class ListingController extends Controller
   }
   public function update(Request $request, Listing $listing)
   {
+    // make sure logged in user is the owner
+    if ($listing->user_id != auth()->id()) {
+      return abort(403, 'Unauthorized action');
+    }
     $formFields = $request->validate([
       'title' => 'required',
       'company' => ['required'],
@@ -75,7 +79,16 @@ class ListingController extends Controller
   }
   public function destroy(Listing $listing)
   {
+    // make sure logged in user is the owner
+    if ($listing->user_id != auth()->id()) {
+      return abort(403, 'Unauthorized action');
+    }
+    
     $listing->delete();
     return redirect('/')->with('message', 'Listing deleted successfully');
+  }
+  public function manage()
+  {
+    return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
   }
 }
